+++ packages/rector-generator/tests/RectorGenerator/Source/StaticRectorRecipeFactory.php
@@ -24,7 +24,7 @@ use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigura
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
 
-    $services->set(SeeSomeClass::class)
+    $services->set(SomeClass::class)
         ->arg('$key', 'value');
 }
 CODE_SAMPLE
@@ -36,7 +36,7 @@ use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigura
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
 
-    $services->set(SeeSomeClass::class)
+    $services->set(SomeClass::class)
         ->call('configure', [[
             '$key' => 'value'
         ]]);
@@ -46,7 +46,7 @@ CODE_SAMPLE
 
         $rectorRecipe->setConfiguration([
             'CLASS_TYPE_TO_METHOD_NAME' => [
-                'SeeSomeClass' => 'configure'
+                'SomeClass' => 'configure'
             ]
         ]);
