+++ rules/doctrine-code-quality/tests/Rector/Class_/MoveCurrentDateTimeDefaultInEntityToConstructorRector/Fixture/fixture.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\DoctrineCodeQuality\Tests\Rector\Property\MoveCurrentDateTimeDefaultInEntityToConstructorRector\Fixture;
+namespace Rector\DoctrineCodeQuality\Tests\Rector\Class_\MoveCurrentDateTimeDefaultInEntityToConstructorRector\Fixture;
 
 use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
@@ -22,7 +22,7 @@ class User
 -----
 <?php
 
-namespace Rector\DoctrineCodeQuality\Tests\Rector\Property\MoveCurrentDateTimeDefaultInEntityToConstructorRector\Fixture;
+namespace Rector\DoctrineCodeQuality\Tests\Rector\Class_\MoveCurrentDateTimeDefaultInEntityToConstructorRector\Fixture;
 
 use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
