+++ rules/downgrade-php74/tests/Rector/Property/DowngradeTypedPropertyRector/SourceNoDocBlock/AnotherClass.php
@@ -1,7 +1,9 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Rector\DowngradePhp74\Tests\Rector\Property\DowngradeTypedPropertyRector\SourceNoDocBlock;
 
-class AnotherClass
+final class AnotherClass
 {
 }
