+++ rules/generic/tests/Rector/ClassMethod/NormalToFluentRector/Fixture/fixture4.php.inc
@@ -1,8 +1,8 @@
 <?php
 
-namespace Rector\Generic\Tests\Rector\MethodCall\NormalToFluentRector\Fixture;
+namespace Rector\Generic\Tests\Rector\ClassMethod\NormalToFluentRector\Fixture;
 
-use Rector\Generic\Tests\Rector\MethodCall\NormalToFluentRector\Source\FluentInterfaceClass;
+use Rector\Generic\Tests\Rector\ClassMethod\NormalToFluentRector\Source\FluentInterfaceClass;
 
 class ActionClass4
 {
@@ -19,9 +19,9 @@ class ActionClass4
 -----
 <?php
 
-namespace Rector\Generic\Tests\Rector\MethodCall\NormalToFluentRector\Fixture;
+namespace Rector\Generic\Tests\Rector\ClassMethod\NormalToFluentRector\Fixture;
 
-use Rector\Generic\Tests\Rector\MethodCall\NormalToFluentRector\Source\FluentInterfaceClass;
+use Rector\Generic\Tests\Rector\ClassMethod\NormalToFluentRector\Source\FluentInterfaceClass;
 
 class ActionClass4
 {
