+++ rules/generic/tests/Rector/ClassMethod/NormalToFluentRector/Source/FluentInterfaceClass.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Generic\Tests\Rector\ClassMethod\NormalToFluentRector\Source;
+
+final class FluentInterfaceClass
+{
+
+}
