--- rules/generic/tests/Rector/MethodCall/NormalToFluentRector/Source/FluentInterfaceClass.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Generic\Tests\Rector\MethodCall\NormalToFluentRector\Source;
-
-final class FluentInterfaceClass
-{
-
-}
