+++ rules/nette-to-symfony/tests/Rector/ClassMethod/RouterListToControllerAnnotationsRector/Fixture/constant_reference_route_to_annotation.php.inc
