+++ rules/nette-to-symfony/tests/Rector/ClassMethod/RouterListToControllerAnnotationsRector/Source/RouteFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\NetteToSymfony\Tests\Rector\ClassMethod\RouterListToControllerAnnotationsRetor\Source;
+namespace Rector\NetteToSymfony\Tests\Rector\ClassMethod\RouterListToControllerAnnotationsRector\Source;
 
 use Nette\Application\Routers\Route;
