+++ rules/php71/tests/Rector/FuncCall/RemoveExtraParametersRector/Fixture/skip_invoke.php.inc
@@ -8,7 +8,7 @@ class SkipInvoke
 {
     public function run($traitMock)
     {
-        $reflectionClass = new ReflectionClass('SeeSomeClass');
+        $reflectionClass = new ReflectionClass('SomeClass');
         $method = $reflectionClass->getMethod('getSetItems');
 
         $method->invoke($traitMock, 'some-param', 4, 5, 6);
