+++ rules/phpunit/tests/Rector/MethodCall/GetMockBuilderGetMockToCreateMockRector/Fixture/fixture.php.inc
@@ -6,7 +6,7 @@ class SomeTest extends \PHPUnit\Framework\TestCase
 {
     public function test()
     {
-        $applicationMock = $this->getMockBuilder('SeeSomeClass')
+        $applicationMock = $this->getMockBuilder('SomeClass')
            ->disableOriginalConstructor()
            ->getMock();
     }
@@ -22,7 +22,7 @@ class SomeTest extends \PHPUnit\Framework\TestCase
 {
     public function test()
     {
-        $applicationMock = $this->createMock('SeeSomeClass');
+        $applicationMock = $this->createMock('SomeClass');
     }
 }
